up_hw2

+++ hw1/reinforce_gae.py
@@ -20,7 +20,7 @@ SavedAction = namedtuple('SavedAction', ['log_prob', 'value'])
 
 # Define a tensorboard writer
 writer = SummaryWriter("./tb_record_3")
-        
+
 class Policy(nn.Module):
     """
         Implement both policy network and the value network in one model
@@ -33,7 +33,7 @@ class Policy(nn.Module):
     """
     def __init__(self):
         super(Policy, self).__init__()
-        
+
         # Extract the dimensionality of state and action spaces
         self.discrete = isinstance(env.action_space, gym.spaces.Discrete)
         self.observation_dim = env.observation_space.shape[0]
@@ -41,7 +41,7 @@ class Policy(nn.Module):
         self.hidden_size = 128
         self.double()
         ########## YOUR CODE HERE (5~10 lines) ##########
-        
+
         self.model=nn.Sequential(
             nn.Linear(self.observation_dim,self.hidden_size),
             nn.LeakyReLU(),
@@ -50,13 +50,13 @@ class Policy(nn.Module):
             nn.Linear(self.hidden_size,self.action_dim),
             nn.Softmax(dim=0),
         )
-        
+
         self.critic=nn.Sequential(
             nn.Linear(self.hidden_size,1),
         )
-        
+
         ########## END OF YOUR CODE ##########
-        
+
         # action & reward memory
         self.saved_actions = []
         self.rewards = []
@@ -64,18 +64,18 @@ class Policy(nn.Module):
     def forward(self, state):
         """
             Forward pass of both policy and value networks
-            - The input is the state, and the outputs are the corresponding 
+            - The input is the state, and the outputs are the corresponding
               action probability distirbution and the state value
             TODO:
                 1. Implement the forward pass for both the action and the state value
         """
-        
+
         ########## YOUR CODE HERE (3~5 lines) ##########
         state =torch.tensor(state)
         out=self.model(state)
         action_prob=self.actor(out)
-        state_value=self.critic(out)       
-    
+        state_value=self.critic(out)
+
         ########## END OF YOUR CODE ##########
 
         return action_prob, state_value
@@ -84,20 +84,20 @@ class Policy(nn.Module):
     def select_action(self, state):
         """
             Select the action given the current state
-            - The input is the state, and the output is the action to apply 
+            - The input is the state, and the output is the action to apply
             (based on the learned stochastic policy)
             TODO:
                 1. Implement the forward pass for both the action and the state value
         """
-        
+
         ########## YOUR CODE HERE (3~5 lines) ##########
-        
+
         action,state_value =self.forward(state)
         m=Categorical(action)
         action=m.sample()
-        
+
         ########## END OF YOUR CODE ##########
-        
+
         # save to action buffer
         self.saved_actions.append(SavedAction(m.log_prob(action), state_value))
 
@@ -112,12 +112,12 @@ class Policy(nn.Module):
                 2. Calculate the policy loss using the policy gradient
                 3. Calculate the value loss using either MSE loss or smooth L1 loss
         """
-        
+
         # Initialize the lists and variables
         R = 0
         saved_actions = self.saved_actions
-        policy_losses = [] 
-        value_losses = [] 
+        policy_losses = []
+        value_losses = []
         returns = []
         val=[]
         act=[]
@@ -125,27 +125,27 @@ class Policy(nn.Module):
         for logprob,state_calue in saved_actions:
             val.append(state_calue[0])
             act.append(logprob)
-            
+
         ########## YOUR CODE HERE (8-15 lines) ##########
         for t in reversed(range(len(self.rewards))):
             R=self.rewards[t]+R*gamma
             returns.insert(0,R)
-        returns=torch.tensor(returns,dtype=torch.float32)
+        returns=torch.tensor(returns,dtype=torch.float64)
         retunrs =(returns -returns.mean())/returns.std()
-        
+
         A =GAE(0.99,0.95,0)
         advantages=A(self.rewards,val,0)
-        
+
         for ad in advantages:
             advan.append(ad.detach())
-            
+
         for t in range(len(self.rewards)):
             policy_losses.append(advan[t]*act[t])
             # value_losses.append(F.huber_loss(returns[t],val[t],delta=1))
             value_losses.append(advan[t].pow(2))
             # print(policy_losses)
             # print(value_losses)
-    
+
         loss= -torch.stack(policy_losses).sum()
         #+torch.stack(value_losses).sum()
         # print(torch.stack(value_losses).mean().item(),end="")
@@ -193,24 +193,24 @@ class GAE:
 def train(lr=0.01):
     """
         Train the model using SGD (via backpropagation)
-        TODO (1): In each episode, 
+        TODO (1): In each episode,
         1. run the policy till the end of the episode and keep the sampled trajectory
         2. update both the policy and the value network at the end of episode
 
-        TODO (2): In each episode, 
+        TODO (2): In each episode,
         1. record all the value you aim to visualize on tensorboard (lr, reward, length, ...)
     """
-    
+
     # Instantiate the policy model and the optimizer
     model = Policy()
     optimizer = optim.Adam(model.parameters(), lr=lr)
-    
+
     # Learning rate scheduler (optional)
     scheduler = Scheduler.StepLR(optimizer, step_size=500, gamma=0.9)
-    
+
     # EWMA reward for tracking the learning progress
     ewma_reward = 0
-    
+
     # run inifinitely many episodes
     for i_episode in count(1):
         # reset environment and episode reward
@@ -220,12 +220,12 @@ def train(lr=0.01):
 
         # Uncomment the following line to use learning rate scheduler
         scheduler.step()
-        
+
         # For each episode, only run 9999 steps to avoid entering infinite loop during the learning process
-        
+
         ########## YOUR CODE HERE (10-15 lines) ##########
         optimizer.zero_grad()
-        
+
         for t in range(9999):
             action = model.select_action(state)
             state, reward, done, _ = env.step(action)
@@ -239,13 +239,13 @@ def train(lr=0.01):
         loss.backward()
         optimizer.step()
         # print(loss)
-        
+
         ########## END OF YOUR CODE ##########
-        
+
         # update EWMA reward and log the results
         ewma_reward = 0.05 * ep_reward + (1 - 0.05) * ewma_reward
         print('Episode {}\tlength: {}\treward: {}\t ewma reward: {}'.format(i_episode, t, ep_reward, ewma_reward))
-        #Try to use Tensorboard to record the behavior of your implementation 
+        #Try to use Tensorboard to record the behavior of your implementation
         ########## YOUR CODE HERE (4-5 lines) ##########
         writer.add_scalar("Vanilla RL Loss",loss,i_episode)
         writer.add_scalar("Vanilla RL Ewma reward",ewma_reward,i_episode)
@@ -267,14 +267,14 @@ def train(lr=0.01):
 def test(name, n_episodes=10):
     """
         Test the learned model (no change needed)
-    """     
+    """
     model = Policy()
-    
+
     model.load_state_dict(torch.load('./preTrained/{}'.format(name)))
-    
+
     render = True
     max_episode_len = 10000
-    
+
     for i_episode in range(1, n_episodes+1):
         state = env.reset()
         running_reward = 0
@@ -288,14 +288,14 @@ def test(name, n_episodes=10):
                 break
         print('Episode {}\tReward: {}'.format(i_episode, running_reward))
     env.close()
-    
+
 
 if __name__ == '__main__':
     # For reproducibility, fix the random seed
-    random_seed = 10  
+    random_seed = 10
     lr = 0.005
     env = gym.make('LunarLander-v2')
-    env.seed(random_seed)  
-    torch.manual_seed(random_seed)  
+    env.seed(random_seed)
+    torch.manual_seed(random_seed)
     train(lr)
     test(f'LunarLander_{lr}.pth')
